requirements.txt
torch
torchvision
torchaudio
transformers
pytorch-lightning
torchmetrics
pandas
numpy
tensorboard
scikit-learn
wandb
ipykernel
ultralytics
diffusers[torch]
gradio
PyYAML
filelock
Pillow
sniffio